// src/features/api/testimonials/messages/testimonial.message.ts
export enum TestimonialsError {
  TESTIMONIALS_NOT_FOUND = 'Testimonial not found',
  TESTIMONIALS_ERROR = 'Oops... There were problems creating or editing the testimonial. Please try again later',
  INTERNAL_SERVER_ERROR = 'Internal server error. Contact with the administrator',
}

export enum TestimonialsOk {
  TESTIMONIALS_CREATED = 'Testimonials created successfully',
  TESTIMONIALS_UPDATED = 'Testimonials updated successfully',
  TESTIMONIALS_REMOVED = 'Testimonials removed successfully',
}